// src/a2aclient/types.ts
export type Part = TextPart | FilePart | DataPart;
/**
 * Mirrors the OpenAPI Security Scheme Object
 * (https://swagger.io/specification/#security-scheme-object)
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SecurityScheme".
 */
export type SecurityScheme =
  | APIKeySecurityScheme
  | HTTPAuthSecurityScheme
  | OAuth2SecurityScheme
  | OpenIdConnectSecurityScheme;
// ...
export type GetTaskResponse = JSONRPCErrorResponse | GetTaskSuccessResponse;
// ...
export type SendMessageResponse = JSONRPCErrorResponse | SendMessageSuccessResponse;
// ...
export interface JSONParseError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32700;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
/**
 * JSON-RPC error indicating the JSON sent is not a valid Request object.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "InvalidRequestError".
 */
export interface InvalidRequestError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32600;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
/**
 * JSON-RPC error indicating the method does not exist or is not available.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "MethodNotFoundError".
 */
export interface MethodNotFoundError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32601;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
/**
 * JSON-RPC error indicating invalid method parameter(s).
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "InvalidParamsError".
 */
export interface InvalidParamsError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32602;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
/**
 * JSON-RPC error indicating an internal JSON-RPC error on the server.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "InternalError".
 */
export interface InternalError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32603;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
// ...
export interface Message {
  /**
   * The context the message is associated with
   */
  contextId?: string;
  /**
   * The URIs of extensions that are present or contributed to this Message.
   */
  extensions?: string[];
  /**
   * Event type
   */
  kind: "message";
  /**
   * Identifier created by the message creator
   */
  messageId: string;
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Message content
   */
  parts: Part[];
  /**
   * List of tasks referenced as context by this message.
   */
  referenceTaskIds?: string[];
  /**
   * Message sender's role
   */
  role: "agent" | "user";
  /**
   * Identifier of task the message is related to
   */
  taskId?: string;
}
// ...
export interface APIKeySecurityScheme {
  /**
   * Description of this security scheme.
   */
  description?: string;
  /**
   * The location of the API key. Valid values are "query", "header", or "cookie".
   */
  in: "cookie" | "header" | "query";
  /**
   * The name of the header, query or cookie parameter to be used.
   */
  name: string;
  type: "apiKey";
}
// ...
export interface AgentExtension {
  /**
   * A description of how this agent uses this extension.
   */
  description?: string;
  /**
   * Optional configuration for the extension.
   */
  params?: {
    [k: string]: unknown;
  };
  /**
   * Whether the client must follow specific requirements of the extension.
   */
  required?: boolean;
  /**
   * The URI of the extension.
   */
  uri: string;
}
/**
 * An AgentCard conveys key information:
 * - Overall details (version, name, description, uses)
 * - Skills: A set of capabilities the agent can perform
 * - Default modalities/content types supported by the agent.
 * - Authentication requirements
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "AgentCard".
 */
export interface AgentCard {
  capabilities: AgentCapabilities1;
  /**
   * The set of interaction modes that the agent supports across all skills. This can be overridden per-skill.
   * Supported media types for input.
   */
  defaultInputModes: string[];
  /**
   * Supported media types for output.
   */
  defaultOutputModes: string[];
  /**
   * A human-readable description of the agent. Used to assist users and
   * other agents in understanding what the agent can do.
   */
  description: string;
  /**
   * A URL to documentation for the agent.
   */
  documentationUrl?: string;
  /**
   * A URL to an icon for the agent.
   */
  iconUrl?: string;
  /**
   * Human readable name of the agent.
   */
  name: string;
  provider?: AgentProvider;
  /**
   * Security requirements for contacting the agent.
   */
  security?: {
    [k: string]: string[];
  }[];
  /**
   * Security scheme details used for authenticating with this agent.
   */
  securitySchemes?: {
    [k: string]: SecurityScheme;
  };
  /**
   * Skills are a unit of capability that an agent can perform.
   */
  skills: AgentSkill[];
  /**
   * true if the agent supports providing an extended agent card when the user is authenticated.
   * Defaults to false if not specified.
   */
  supportsAuthenticatedExtendedCard?: boolean;
  /**
   * A URL to the address the agent is hosted at.
   */
  url: string;
  /**
   * The version of the agent - format is up to the provider.
   */
  version: string;
}
/**
 * Optional capabilities supported by the agent.
 */
export interface AgentCapabilities1 {
  /**
   * extensions supported by this agent.
   */
  extensions?: AgentExtension[];
  /**
   * true if the agent can notify updates to client.
   */
  pushNotifications?: boolean;
  /**
   * true if the agent exposes status change history for tasks.
   */
  stateTransitionHistory?: boolean;
  /**
   * true if the agent supports SSE.
   */
  streaming?: boolean;
}
/**
 * The service provider of the agent
 */
export interface AgentProvider {
  /**
   * Agent provider's organization name.
   */
  organization: string;
  /**
   * Agent provider's URL.
   */
  url: string;
}
/**
 * HTTP Authentication security scheme.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "HTTPAuthSecurityScheme".
 */
export interface HTTPAuthSecurityScheme {
  /**
   * A hint to the client to identify how the bearer token is formatted. Bearer tokens are usually
   * generated by an authorization server, so this information is primarily for documentation
   * purposes.
   */
  bearerFormat?: string;
  /**
   * Description of this security scheme.
   */
  description?: string;
  /**
   * The name of the HTTP Authentication scheme to be used in the Authorization header as defined
   * in RFC7235. The values used SHOULD be registered in the IANA Authentication Scheme registry.
   * The value is case-insensitive, as defined in RFC7235.
   */
  scheme: string;
  type: "http";
}
/**
 * OAuth2.0 security scheme configuration.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "OAuth2SecurityScheme".
 */
export interface OAuth2SecurityScheme {
  /**
   * Description of this security scheme.
   */
  description?: string;
  flows: OAuthFlows;
  type: "oauth2";
}
// ...
export interface OpenIdConnectSecurityScheme {
  /**
   * Description of this security scheme.
   */
  description?: string;
  /**
   * Well-known URL to discover the [[OpenID-Connect-Discovery]] provider metadata.
   */
  openIdConnectUrl: string;
  type: "openIdConnect";
}
/**
 * Represents a unit of capability that an agent can perform.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "AgentSkill".
 */
export interface AgentSkill {
  /**
   * Description of the skill - will be used by the client or a human
   * as a hint to understand what the skill does.
   */
  description: string;
  /**
   * The set of example scenarios that the skill can perform.
   * Will be used by the client as a hint to understand how the skill can be used.
   */
  examples?: string[];
  /**
   * Unique identifier for the agent's skill.
   */
  id: string;
  /**
   * The set of interaction modes that the skill supports
   * (if different than the default).
   * Supported media types for input.
   */
  inputModes?: string[];
  /**
   * Human readable name of the skill.
   */
  name: string;
  /**
   * Supported media types for output.
   */
  outputModes?: string[];
  /**
   * Set of tagwords describing classes of capabilities for this specific skill.
   */
  tags: string[];
}
// ...
export interface Artifact {
  /**
   * Unique identifier for the artifact.
   */
  artifactId: string;
  /**
   * Optional description for the artifact.
   */
  description?: string;
  /**
   * The URIs of extensions that are present or contributed to this Artifact.
   */
  extensions?: string[];
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Optional name for the artifact.
   */
  name?: string;
  /**
   * Artifact parts.
   */
  parts: Part[];
}
// ...
export interface JSONRPCErrorResponse {
  error:
    | JSONRPCError
    | JSONParseError
    | InvalidRequestError
    | MethodNotFoundError
    | InvalidParamsError
    | InternalError
    | TaskNotFoundError
    | TaskNotCancelableError
    | PushNotificationNotSupportedError
    | UnsupportedOperationError
    | ContentTypeNotSupportedError
    | InvalidAgentResponseError;
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
}
/**
 * Represents a JSON-RPC 2.0 Error object.
 * This is typically included in a JSONRPCErrorResponse when an error occurs.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "JSONRPCError".
 */
export interface JSONRPCError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: number;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
// ...
export interface Task {
  /**
   * Collection of artifacts created by the agent.
   */
  artifacts?: Artifact[];
  /**
   * Server-generated id for contextual alignment across interactions
   */
  contextId: string;
  history?: Message1[];
  /**
   * Unique identifier for the task
   */
  id: string;
  /**
   * Event type
   */
  kind: "task";
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  status: TaskStatus;
}
/**
 * Represents a single message exchanged between user and agent.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "Message".
 */
export interface Message1 {
  /**
   * The context the message is associated with
   */
  contextId?: string;
  /**
   * The URIs of extensions that are present or contributed to this Message.
   */
  extensions?: string[];
  /**
   * Event type
   */
  kind: "message";
  /**
   * Identifier created by the message creator
   */
  messageId: string;
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Message content
   */
  parts: Part[];
  /**
   * List of tasks referenced as context by this message.
   */
  referenceTaskIds?: string[];
  /**
   * Message sender's role
   */
  role: "agent" | "user";
  /**
   * Identifier of task the message is related to
   */
  taskId?: string;
}
/**
 * Current status of the task
 */
export interface TaskStatus {
  message?: Message2;
  state: TaskState;
  /**
   * ISO 8601 datetime string when the status was recorded.
   */
  timestamp?: string;
}
// ...
export interface GetTaskSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  result: Task1;
}
// ...
export interface SendMessageSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  /**
   * The result object on success
   */
  result: Task2 | Message1;
}
// ...
export interface TaskStatusUpdateEvent {
  /**
   * The context the task is associated with
   */
  contextId: string;
  /**
   * Indicates the end of the event stream
   */
  final: boolean;
  /**
   * Event type
   */
  kind: "status-update";
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  status: TaskStatus1;
  /**
   * Task id
   */
  taskId: string;
}
/**
 * Current status of the task
 */
export interface TaskStatus1 {
  message?: Message2;
  state: TaskState;
  /**
   * ISO 8601 datetime string when the status was recorded.
   */
  timestamp?: string;
}